transmission-received/requirements.txt
PySSTV
pillow
python-zbase32
qrcode